COMMERCE-5657 dispatch-web - extract duplicated code
